Make Debug_Log* functions to crossplatform
